test/requirements.txt
python-etcd==0.3.2
requests==2.3.0
nose==1.3.3